+++ framework/core/src/services/index.ts
@@ -1,4 +1,4 @@
-export * from '../../../common/src/getEnvVar';
+export { getEnvVar } from '@forklaunch/common';
 export * from '../environment';
 export { createConfigInjector } from './configInjector';
 export type { ConfigInjector, ValidConfigInjector } from './configInjector';